% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpsa_mentalhealth}
\alias{hpsa_mentalhealth}
\title{HPSA Mental Health}
\format{
An object of class \code{data.frame} with 149 rows and 16 columns.
}
\source{
\url{https://data.chhs.ca.gov/dataset/42406266-89d3-4d04-9c45-587de774a1d5/resource/8d10c854-9cf8-459d-8d30-5b1b60eeb845/download/hpsa_mentalhealth.csv}
}
\usage{
hpsa_mentalhealth
}
\description{
More detailed documentation will go here
}
\keyword{datasets}
