% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpsa_primarycare}
\alias{hpsa_primarycare}
\title{HPSA Primary Care}
\format{
An object of class \code{data.frame} with 230 rows and 17 columns.
}
\source{
\url{https://data.chhs.ca.gov/dataset/42406266-89d3-4d04-9c45-587de774a1d5/resource/e6b6e925-f2b5-4a5b-ab5e-515ef55294b3/download/hpsa_primarycare.csv}
}
\usage{
hpsa_primarycare
}
\description{
More detailed documentation will go here
}
\keyword{datasets}
