% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_expectancy}
\alias{life_expectancy}
\title{Life Expectancy at Birth for U.S. States and Census Tracts, 2010-2015}
\format{
An object of class \code{data.frame} with 73121 rows and 6 columns.
}
\source{
\url{https://www.cdc.gov/nchs/data-visualization/life-expectancy/index.html}
}
\usage{
life_expectancy
}
\description{
Life Expectancy at Birth for U.S. States and Census Tracts, 2010-2015
}
\keyword{datasets}
