% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebooks.R
\name{make_codebook}
\alias{make_codebook}
\title{Codebook Maker}
\usage{
make_codebook(
  input_df,
  return_df = TRUE,
  format = NULL,
  escape = TRUE,
  extra_vars = NULL,
  extra_key = NULL,
  ...
)
}
\arguments{
\item{input_df}{A \code{data.frame} that will be summarized as a codebook.}

\item{return_df}{A boolean set by default to \code{TRUE}, if \code{FALSE}
it will return the table as HTML or LaTeX.}

\item{format}{A string set to \code{NULL} by default. If \code{NULL} and
the \code{return_df} parameter is set to \code{FALSE}, this argument may be
set to \code{"html"} or \code{"latex"} to determine the output of the table.}

\item{escape}{A boolean; whether to escape special characters when producing
HTML or LaTeX tables. When \code{escape = FALSE}, you have to make sure that
special characters will not trigger syntax errors in LaTeX or HTML.}

\item{extra_vars}{A \code{data.frame} of variables that wil be left joined.}

\item{extra_key}{A character; specifying the column used for merging the \code{extra_vars} table.}

\item{...}{Other arguments (see Examples and References).}
}
\description{
Codebook Maker
}
\examples{
\dontrun{
# Create a test data.frame
test <- data.frame("Incident_Type" = c("WSP", "WMCI", "WHF", "WWC", "WHCC", 
                                        "WSP", "WMCI", "WHF", "WWC", "WHCC"),
 "n" = seq(1, 10),
 "n2" = as.double( seq(11, 20) ),
 "string" = c(LETTERS[1:10]),
 "test2" = c("A", "A", "B", "B", "C", "C", "D", "E", "F", "A"),
 "factor" = as.factor(rep(letters[1:5], 2)),
 "test" = c(rep(TRUE, 6), FALSE, NA, NA, NA)
 )

# Return HTML table.
codemaker(test, return_df =  FALSE, format = "html", 
latex_options = "striped", escape = FALSE, extra_vars = added_vars, 
extra_key = "var_name")

}


}
