% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebooks.R
\name{modefinder}
\alias{modefinder}
\title{Find Vector Mode}
\usage{
modefinder(v, na.rm = TRUE)
}
\arguments{
\item{v}{is an atomic vector.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
Find Vector Mode
}
\examples{
\dontrun{
ex <- as.Date(c("2020-01-01", "2020-01-01", "2020-01-01", "2020-01-01", "2020-01-02", "2020-01-02", "2020-01-03", "2020-01-03"))
modefinder(ex)
}

}
\keyword{internal}
