% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_state_and_county}
\alias{saipe_state_and_county}
\title{Small Area Income and Poverty Estimates (SAIPE) State and County Estimates for 2021}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3195 rows and 11 columns.
}
\source{
\url{https://www.census.gov/data/datasets/2021/demo/saipe/2021-state-and-county.html}
}
\usage{
saipe_state_and_county
}
\description{
Small Area Income and Poverty Estimates (SAIPE) State and County Estimates for 2021
}
\keyword{datasets}
