% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_sngl_var}
\alias{visualize_sngl_var}
\title{Visualize Single Variable}
\usage{
visualize_sngl_var(
  df,
  var,
  xlab,
  ylab = NULL,
  theme = NULL,
  widget = FALSE,
  flip = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame from which a variable will be visualized.}

\item{var}{A character representing the name of a single variable of a given
data frame.}

\item{xlab}{An required character representing the label that will be places
on the x-axis.}

\item{ylab}{An optional character representing the label that will be places
on the y-axis.}

\item{theme}{A ggplot theme used to tweak the display of the graph.}

\item{widget}{A boolean which defaults to \code{FALSE}. When \code{TRUE}, the graph is
returned as an interactive HTML widget powered by the {plotly} library.}

\item{flip}{A boolean which defaults to \code{FALSE}. When \code{TRUE}, the graph is
turned on its axis.}

\item{...}{Other arguments passed on to the geom functions.}
}
\description{
An opinionated function that delivers single varible descriptive graphs based
on the variable's class.
}
\examples{
\dontrun{
d <- data.frame(
  "char" = rep(LETTERS[1:5], 10),
  "ints" = rep(c(1L, 2L), 25)
 )
 
visualize_sngl_var(d, "ints", widget = TRUE, xlab = "Integers",
   ylab ="Count")
}


}
